# swagger-spec.json
{"openapi":"3.0.0","paths":{"/tracts/all":{"get":{"operationId":"CensusTractController_getAll","parameters":[],"responses":{"200":{"description":""}},"tags":["CensusTract"]}},"/tracts/within-boundary":{"post":{"operationId":"CensusTractController_findWithinBoundary","parameters":[],"responses":{"201":{"description":""}},"tags":["CensusTract"]}},"/tracts/income":{"post":{"operationId":"CensusTractController_findIncome","parameters":[],"responses":{"201":{"description":""}},"tags":["CensusTract"]}},"/tracts/rent":{"post":{"operationId":"CensusTractController_findRent","parameters":[],"responses":{"201":{"description":""}},"tags":["CensusTract"]}},"/tracts/search/city/{searchQuery}":{"get":{"operationId":"CensusTractController_searchCity","parameters":[{"name":"searchQuery","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["CensusTract"]}},"/location/search/city/{searchQuery}":{"get":{"operationId":"LocationController_searchCity","parameters":[{"name":"searchQuery","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Location"]}},"/block/within-boundary":{"post":{"operationId":"CensusBlockGroupController_getWithinBoundary","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BoundaryDto"}}}},"responses":{"201":{"description":""}},"tags":["CensusBlockGroup"]}},"/block/sex-by-age":{"post":{"operationId":"CensusBlockGroupController_getSexByAge","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BoundaryDto"}}}},"responses":{"201":{"description":""}},"tags":["CensusBlockGroup"]}},"/block/income-per-capita":{"post":{"operationId":"CensusBlockGroupController_getIncomePerCapita","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BoundaryDto"}}}},"responses":{"201":{"description":""}},"tags":["CensusBlockGroup"]}}},"info":{"title":"CensusData","description":"Census Data API","version":"1.0","contact":{}},"tags":[{"name":"CensusData","description":""}],"servers":[],"components":{"schemas":{"BoundaryDto":{"type":"object","properties":{}}}}}